backend/requirements.txt
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
docstring_parser==0.17.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
multidict==6.7.0
packaging==25.0
pillow==12.0.0
postgrest==2.25.1
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
qrcode==8.2
realtime==2.25.1
sniffio==1.3.1
sqlparse==0.5.4
storage3==2.25.1
StrEnum==0.4.15
supabase==2.25.1
supabase-auth==2.25.1
supabase-functions==2.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
websockets==15.0.1
yarl==1.22.0
